リダイレクション周りの関数を cmd_redirection.c に入れた
