+++ test/Makefile
@@ -4,7 +4,7 @@ LIBFT_LIB = ../libft/libft.a
 SRCS = 	../env.c ../env_setter.c ../path.c															\
 		../lexer1.c ../lexer2.c																		\
 		../parse1.c ../parse2.c ../parse_utils.c ../parse_utils2.c									\
-		../exec.c ../cmd_exec_command.c ../cmd_exec_commands.c ../cmd_pipe.c						\
+		../exec.c ../cmd_exec_command.c ../cmd_exec_commands.c ../cmd_redirection.c ../cmd_pipe.c	\
 		../cmd_cmd_invocation.c ../cmd_cmd_invocation2.c ../convert_ast2cmdinvo.c					\
 		../expand_env_var.c ../split_expanded_str.c ../string_node2string.c ../expand_string_node.c	\
 		../cmd_status.c ../signal.c ../cmd_status.c ../signal.c ../minishell_error_msg.c			\
